refactor(FE): 모달 상태 관리 로직 최적화
- `ReviewDetailModal`에서 URL 쿼리 파라미터에 따라 모달 상태 관리하도록 변경
- `Feed` 컴포넌트에서 불필요한 코드 제거
